Made Admin Functional Component. Created All User Class Component
