+++ client/components/AdminPage.js
@@ -1,36 +1,15 @@
-import React, {Component} from 'react'
-import {connect} from 'react-redux'
+import React from 'react'
 import {Link} from 'react-router-dom'
-import {getAllUsers} from '../store/user'
 
-export class AdminPage extends Component {
-  componentDidMount() {
-    this.props.getAllUsers()
-  }
-  render() {
-    console.log('Users', this.props.users)
-    return (
-      <div className="single-product-div">
-        <h1>Admin Page</h1>
-        <h2>All site Users</h2>
-        {this.props.users.map(user => (
-          <Link key={user.id} to={`/user/${user.id}`}>
-            <div>
-              <h3>{user.name}</h3>
-            </div>
-          </Link>
-        ))}
-      </div>
-    )
-  }
+export const AdminPage = () => {
+  return (
+    <div className="single-product-div">
+      <h1>Admin Page</h1>
+      <Link to="/allUsers">All Users</Link>
+      <br />
+      <Link to="/products">All Products</Link>
+    </div>
+  )
 }
 
-const mapStateToProps = state => ({
-  users: state.user.allUsers
-})
-
-const mapDispatchToProps = dispatch => ({
-  getAllUsers: () => dispatch(getAllUsers())
-})
-
-export default connect(mapStateToProps, mapDispatchToProps)(AdminPage)
+export default AdminPage
