+++ client/components/index.js
@@ -8,6 +8,7 @@ export {default as UserHome} from './user-home'
 export {default as Home} from './Home'
 export {default as ProductList} from './ProductList'
 export {default as Product} from './Product'
+export {default as AllUsers} from './AllUsers'
 export {default as SingleUser} from './SingleUser'
 export {default as Cart} from './Cart'
 export {default as AdminPage} from './AdminPage'
